+++ src/datagrid/datagrid.component.jsx
@@ -781,8 +781,14 @@ export default class DataGrid extends React.PureComponent {
             }
             break;
           case 'select': {
+            const { intl } = this.props;
             const selectOptions = col.selectComponentOptions ||
               this.props.selectComponentOptions.get(column.columnKey);
+            const selectTranslations = col.selectComponentTranslations ||
+              {
+                placeholder: intl.formatMessage({ id: 'Grid.FloatingSelect.Select' }),
+                noResultsText: intl.formatMessage({ id: 'Grid.FloatingSelect.NoResults' }),
+              };
             if (this.props.inlineEdit) {
               if (!column.cellEdit) {
                 column.cellEdit = rowIndex => (
@@ -809,6 +815,7 @@ export default class DataGrid extends React.PureComponent {
                     {...col.editComponentProps}
                     disabled={this.getComponentDisabledState(rowIndex, col, 'edit')}
                     tabIndex={tabIndex}
+                    {...selectTranslations}
                   />
                 );
               }
@@ -836,6 +843,7 @@ export default class DataGrid extends React.PureComponent {
                     {...col.createComponentProps}
                     disabled={this.getComponentDisabledState(rowIndex, col, 'create')}
                     tabIndex={tabIndex}
+                    {...selectTranslations}
                   />
                 );
               }
@@ -860,6 +868,7 @@ export default class DataGrid extends React.PureComponent {
                     }}
                     {...col.filterComponentProps}
                     tabIndex={tabIndex}
+                    {...selectTranslations}
                   />
                 );
               }