+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## <next>
 * Added option selectComponentTranslations to customize react-selects default placeholder and noResultsTexts. By default uses the Grids translations.
+* Updated `@opuscapita/react-datetime` to version 2.2.3.
+
 ## 5.1.8
 * Upgraded `@opuscapita/react-select-order-list` to the latest version.